image_and_video/app/requirements.txt
streamlit>=1.22.0
Pillow>=9.0.0
boto3>=1.26.0
numpy>=1.20.0